# docs/swagger.yaml
definitions:
  generic-integration-platform_internal_application_dto.EndpointRequestDTO:
    properties:
      headers:
        description: Additional headers (optional)
        type: string
      method:
        description: HTTP method (e.g., GET, POST)
        type: string
      name:
        description: Name of the endpoint
        type: string
      path:
        description: Path of the endpoint
        type: string
    required:
    - method
    - name
    - path
    type: object
  generic-integration-platform_internal_application_dto.EndpointResponseDTO:
    properties:
      method:
        description: HTTP method (e.g., GET, POST)
        type: string
      name:
        description: Name of the endpoint
        type: string
      path:
        description: Path of the endpoint
        type: string
    type: object
  generic-integration-platform_internal_application_dto.FlowDTO:
    properties:
      id:
        description: Unique identifier for the flow
        type: string
      name:
        description: Name of the flow
        type: string
      steps:
        description: List of steps in the flow
        items:
          $ref: '#/definitions/generic-integration-platform_internal_application_dto.StepDTO'
        type: array
    type: object
  generic-integration-platform_internal_application_dto.IntegrationRequestDTO:
    properties:
      auth_token:
        description: The authentication token (optional)
        type: string
      auth_type:
        description: Type of authentication (e.g., Bearer, Basic)
        type: string
      base_url:
        description: Base URL for API requests
        type: string
      currency:
        description: Currency for transactions
        type: string
      endpoints:
        description: List of endpoints associated with this integration
        items:
          $ref: '#/definitions/generic-integration-platform_internal_application_dto.EndpointRequestDTO'
        type: array
      name:
        description: Name of the integration
        type: string
      type:
        description: Type of integration (e.g., REST, gRPC)
        type: string
    required:
    - base_url
    - currency
    - endpoints
    - name
    - type
    type: object
  generic-integration-platform_internal_application_dto.IntegrationResponseDTO:
    properties:
      auth_type:
        description: Type of authentication (e.g., Bearer, Basic)
        type: string
      base_url:
        description: Base URL for API requests
        type: string
      currency:
        description: Currency for transactions
        type: string
      endpoints:
        description: List of endpoints associated with this integration
        items:
          $ref: '#/definitions/generic-integration-platform_internal_application_dto.EndpointResponseDTO'
        type: array
      id:
        description: Unique identifier for the integration
        type: string
      name:
        description: Name of the integration
        type: string
      type:
        description: Type of integration (e.g., REST, gRPC)
        type: string
    type: object
  generic-integration-platform_internal_application_dto.StepDTO:
    properties:
      action:
        description: Action to be performed in the step
        type: string
      integration_id:
        description: ID of the associated integration
        type: string
      params:
        additionalProperties:
          type: string
        description: Parameters for the step
        type: object
    type: object
  generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO:
    properties:
      message:
        description: Error message to describe the issue
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an Generic Integration Platform API that processes multiple
    integration and integration flows.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Generic Integration Platform API
  version: "1.0"
paths:
  /flows:
    get:
      description: Retrieve a list of all flows
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/generic-integration-platform_internal_application_dto.FlowDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Get all flows
      tags:
      - Flows
    post:
      consumes:
      - application/json
      description: Create a new flow by providing flow details
      parameters:
      - description: Flow Data
        in: body
        name: flow
        required: true
        schema:
          $ref: '#/definitions/generic-integration-platform_internal_application_dto.FlowDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_application_dto.FlowDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Create a new flow
      tags:
      - Flows
  /flows/{id}:
    delete:
      description: Remove a specific flow by its ID
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Delete a flow by ID
      tags:
      - Flows
    get:
      description: Retrieve details of a specific flow by ID
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_application_dto.FlowDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Get flow details by ID
      tags:
      - Flows
    put:
      consumes:
      - application/json
      description: Update an existing flow by providing its ID and updated details
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Flow Data
        in: body
        name: flow
        required: true
        schema:
          $ref: '#/definitions/generic-integration-platform_internal_application_dto.FlowDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_application_dto.FlowDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Update a flow
      tags:
      - Flows
  /flows/{id}/execute:
    post:
      description: Execute a specific flow by its ID
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_application_dto.FlowDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Execute a flow by ID
      tags:
      - Flows
  /integrations:
    get:
      description: Get a list of all integrations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/generic-integration-platform_internal_application_dto.IntegrationResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: List all integrations
      tags:
      - Integrations
    post:
      consumes:
      - application/json
      description: Create a new integration with the provided details
      parameters:
      - description: Integration data
        in: body
        name: integration
        required: true
        schema:
          $ref: '#/definitions/generic-integration-platform_internal_application_dto.IntegrationRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_application_dto.IntegrationResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Create a new integration
      tags:
      - Integrations
  /integrations/{id}:
    delete:
      description: Remove a specific integration by its ID
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Delete an integration by ID
      tags:
      - Integrations
    get:
      description: Retrieve details of a specific integration by its ID
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_application_dto.IntegrationResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Get a specific integration by ID
      tags:
      - Integrations
    put:
      consumes:
      - application/json
      description: Update the details of an existing integration
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated integration data
        in: body
        name: integration
        required: true
        schema:
          $ref: '#/definitions/generic-integration-platform_internal_application_dto.IntegrationRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_application_dto.IntegrationResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/generic-integration-platform_internal_infra_http_dto.ErrorResponseDTO'
      summary: Update an existing integration by ID
      tags:
      - Integrations
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: 'To access protected routes, add the x-api-key header with your API
      key. Example: "x-api-key: your-api-key"'
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"